+++ variables.tf
@@ -266,14 +266,22 @@ variable "os_disk_caching" {
 variable "log_analytics_workspace_guid" {
   description = "GUID of the Log Analytics Workspace to link with"
   type        = string
+  default     = null
 }
 
 variable "log_analytics_workspace_key" {
   description = "Access key of the Log Analytics Workspace to link with"
   type        = string
+  default     = null
 }
 
 variable "azure_monitor_data_collection_rule_id" {
   description = "Data Collection Rule ID from Azure Monitor for metrics and logs collection"
   type        = string
+}
+
+variable "log_analytics_extension_enable" {
+  description = "Deploy Log Analytics VM extension - depending of OS (cf. https://docs.microsoft.com/fr-fr/azure/azure-monitor/agents/agents-overview#linux)"
+  type        = bool
+  default     = true
 }