+++ r-diagnostics.tf
@@ -4,27 +4,22 @@ module "vm_logs" {
   source  = "claranet/vm-logs/azurerm"
   version = "3.0.0"
 
-  publisher                  = "Microsoft.Azure.Monitor"
-  type                       = "AzureMonitorLinuxAgent"
-  type_handler_version       = "1.0"
-  auto_upgrade_minor_version = "true"
+  location    = var.location
+  client_name = var.client_name
+  environment = var.environment
+  stack       = var.stack
 
-  virtual_machine_id = azurerm_linux_virtual_machine.vm.id
-}
+  resource_group_name = var.resource_group_name
 
-resource "null_resource" "azure_monitor_link" {
-  provisioner "local-exec" {
-    command = <<EOC
-      az rest --subscription ${data.azurerm_client_config.current.subscription_id} \
-              --method PUT \
-              --url https://management.azure.com${azurerm_linux_virtual_machine.vm.id}/providers/Microsoft.Insights/dataCollectionRuleAssociations/${azurerm_linux_virtual_machine.vm.name}-dcrassociation?api-version=2019-11-01-preview \
-              --body '{"properties":{"dataCollectionRuleId": "${var.azure_monitor_data_collection_rule_id}"}}'
-EOC
-  }
+  diagnostics_storage_account_name      = var.diagnostics_storage_account_name
+  diagnostics_storage_account_sas_token = var.diagnostics_storage_account_sas_token
 
-  triggers = {
-    dcr_id = var.azure_monitor_data_collection_rule_id
-    vm_id  = azurerm_linux_virtual_machine.vm.id
+  vm_ids = [azurerm_linux_virtual_machine.vm.id]
+
+  tags = {
+    environment = var.environment
+    stack       = var.stack
+    vm_name     = azurerm_linux_virtual_machine.vm.name
   }
 }