+++ variables.tf
@@ -303,3 +303,15 @@ variable "log_analytics_agent_version" {
   type        = string
   default     = "1.13"
 }
+
+variable "identity" {
+  description = "Map with identity block informations as described here https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/linux_virtual_machine#identity"
+  type = object({
+    type         = string
+    identity_ids = list(string)
+  })
+  default = {
+    type         = "SystemAssigned"
+    identity_ids = []
+  }
+}