+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Changed
 # v4.1.1 - 2021-02-26
 
 Fixed
-  * AZ-451: set `zone_id` as optional
+  * AZ-451: set `zone_id` as optional 
 
 # v4.1.0 - 2021-01-08