+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Changed
   * AZ-302: Bump Log Analytics version to latest and allow override
   * AZ-572: Revamp examples and improve CI
 
+Added
+  * AZ-576: Allow deploying Spot Instances
+
 # v4.1.2 - 2021-08-20
 
 Changed